Yarim gunu ucrete yarim gun olarak yansit
Yarim gun tam sayiliyordu; calisilmayan yarim gun icin ucret odeniyor,
puantaj cetveli ile bordro da birbirini tutmuyordu.

- gun_esdegeri() tek kaynak: tam 1, yarim 0,5; devamsiz ve izinli
  sayilmaz. Puantaj cetveli, puantaj ozeti, bordro ozeti ve aylik ozet
  ayni yardimciyi kullaniyor — dordu ayri ayri sayiyordu
- Bordro.calisilan_gun Integer -> Numeric(5,2); 23,5 gun saklanabiliyor
- Gun sayilari arayuzde Turkce ondalikla yaziliyor (23,5 gun)
- Bordro formundaki gun alani 0,5 adimli; sabit 22 varsayilani kalkti

Mevcut iki taslak bordro yeni kurala gore yeniden hesaplandi:
Kemal Yilmaz 24 -> 23,5 gun (brut 60.000 -> 58.750).

226 test geciyor (218 -> 226).

+++ backend/models.py
@@ -135,7 +135,7 @@ class Bordro(Base):
     gelir_vergisi   = Column(Numeric(12, 2), nullable=False)
     damga_vergisi   = Column(Numeric(12, 2), nullable=False)   # %0.759
     net_maas        = Column(Numeric(12, 2), nullable=False)
-    calisilan_gun   = Column(Integer, default=0)
+    calisilan_gun   = Column(Numeric(5, 2), default=0)   # yarım günler 0,5
     fazla_mesai_saat = Column(Numeric(6, 2), default=0)
     durum           = Column(Enum(BordroDurum), default=BordroDurum.taslak)
     notlar          = Column(String, nullable=True)
